add abstraction for tx records
